requirements.txt
aiogram==3.27.0
python-dotenv==1.2.2